+++ s/transports/post-message/portals.ts
@@ -1,21 +1,29 @@
 
 import {PortalChannel} from "./types.js"
+import {JsonRpc} from "../../comms/json-rpc.js"
 
 export abstract class BasePortal {
 	abstract channel: PortalChannel
-	abstract send(message: any, transfer?: Transferable[]): void
+	abstract sendRequest(message: JsonRpc.Requestish, transfer: Transferable[] | undefined, done: Promise<JsonRpc.Respondish | null>): void
+	abstract sendResponse(message: JsonRpc.Respondish, transfer: Transferable[] | undefined): void
 }
 
 export class WindowPortal extends BasePortal {
 	constructor(public channel: Window, public targetOrigin: string) { super() }
-	send(message: any, transfer?: Transferable[]) {
+	sendRequest(message: JsonRpc.Requestish, transfer: Transferable[] | undefined) {
+		this.channel.postMessage(message, this.targetOrigin, transfer)
+	}
+	sendResponse(message: JsonRpc.Respondish, transfer: Transferable[] | undefined) {
 		this.channel.postMessage(message, this.targetOrigin, transfer)
 	}
 }
 
 export class BroadcastPortal extends BasePortal {
 	constructor(public channel: BroadcastChannel) { super() }
-	send(message: any, _transfer?: Transferable[]) {
+	sendRequest(message: JsonRpc.Requestish, _transfer: Transferable[] | undefined) {
+		this.channel.postMessage(message)
+	}
+	sendResponse(message: JsonRpc.Respondish, _transfer: Transferable[] | undefined) {
 		this.channel.postMessage(message)
 	}
 }
@@ -23,11 +31,15 @@ export class BroadcastPortal extends BasePortal {
 export class MessagePortal extends BasePortal {
 	constructor(
 		public channel: {
-			postMessage(message: any, transfer?: Transferable[]): void
+			postMessage(message: JsonRpc.Bidirectional, transfer?: Transferable[]): void
 		} & PortalChannel
 	) { super() }
 
-	send(message: any, transfer?: Transferable[]) {
+	sendRequest(message: JsonRpc.Requestish, transfer?: Transferable[]) {
+		this.channel.postMessage(message, transfer)
+	}
+
+	sendResponse(message: JsonRpc.Respondish, transfer?: Transferable[]) {
 		this.channel.postMessage(message, transfer)
 	}
 }